requirements.txt
fastapi
uvicorn
notion-client
pydantic-settings